Using payload for PickAllowableFields
