+++ package.json
@@ -1,6 +1,6 @@
 {
   "name": "@cmsgov/wi-validation",
-  "version": "0.0.4",
+  "version": "0.0.5",
   "description": "CMS Web Interface Client and API Validation",
   "keywords": [
     "validation",
